final-project/requirements.txt
bitarray==1.0.1
Pillow==6.1.0
